refactor: Add cover image to Test blog post
Update the Test.md file in the src/content/blogs directory to include a cover image. The cover image URL is set to /images/projects/codechefbcetchapter.png and the alt text is set to "just a test image". This change enhances the visual appeal of the blog post and improves the overall user experience.