refactor: Update BlogPost layout to use smaller calendar icon
This commit updates the BlogPost layout in the src/layouts/BlogPost.astro file to use a smaller calendar icon. The previous icon size was too large and didn't align well with the text. By reducing the size of the calendar icon, the layout now looks more visually balanced and improves the overall design of the blog post page.